+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 ## Change log
 ----------------------
 
+Version 4.1
+-------------
+
+ADDED:
+
+- gradle as build system
+- new method for check if a row is empty created
+- new method get the cell value as object created
+- new class ReadExcelExtensions that reads workbook(excel) files
+
+CHANGED:
+
+- update of dependency poi to new version 5.0.0
+- update of dependency lombok to new version 1.18.20
+- update of dependency file-worker to new version 5.10
+- update of test dependency silly-io to new version 1.6
+- update of test dependency testng to new version 7.4.0
+- removed unused dependencies
+
 Version 4.0.1
 -------------